+++ src/components/PostForm/Inputs/ExpandableInput/DeletableImage/DeletableImage.style.ts
