+++ src/components/PostForm/Inputs/ExpandableInput/ExpandableInput.style.ts
