+++ src/components/PostForm/Inputs/FeeInput/FeeInput.tsx
