+++ src/components/PostForm/Inputs/MinMaxInput/MinMaxInput.style.ts
