+++ src/components/PostForm/Inputs/MinMaxInput/MinMaxInput.tsx
