+++ src/components/PostForm/Inputs/ThumbnailInput/ThumbnailInput.style.ts
