+++ src/components/PostForm/Inputs/ThumbnailInput/ThumbnailInput.tsx
