--- src/pages/PostPage/components/index.ts
@@ -1,1 +1,0 @@
-export { default as PostForm } from "./postForm/PostForm";
