--- src/pages/PostPage/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useGeocoder } from "./useGeocoder";
